requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
Django==2.2.3
django-templated-mail==1.1.1
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
docopt==0.6.2
idna==2.8
oauthlib==3.1.0
Pillow==6.1.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
transliterate==1.10.2
urllib3==1.25.6
yarg==0.1.9
